+++ app/src/main/java/org/techtown/soptseminar/RepositoryFragment.kt
@@ -32,7 +32,9 @@ class RepositoryFragment : Fragment() {
         repositoryAdapter = RepositoryAdapter()
         with(binding) {
             rvRepository.adapter = repositoryAdapter
-            rvRepository.addItemDecoration(MyDecoration(this@RepositoryFragment))
+            rvRepository.addItemDecoration(ItemDecoration(resources.getDimensionPixelOffset(R.dimen.margin_15), 2))
+            // resources : 옛날이름 getResources 메서드
+            // getDimesionPixelOffset메서드 : getDimension메서드와 비슷, (부동소수점을 버림 -> int형)
         }
 
         repositoryAdapter.repositoryList.addAll(
@@ -41,7 +43,7 @@ class RepositoryFragment : Fragment() {
                 RepositoryData("today_junelog", "파이썬 + 자료구조/알고리즘 레포지토리"),
                 RepositoryData("Android_Bymurjune", "자바 + 코틀린 + 안스 레포지토리"),
                 RepositoryData("Hott6/Hot-JunWon", "30기 Sopt 안드로이드 과제 레포지토리"),
-                RepositoryData(" murjune.github.io ", "기술블로"),
+                RepositoryData(" murjune.github.io ", "기술블로그"),
                 RepositoryData(
                     "OpportunityPod/hackerrank-interview-preparation-kit",
                     "Hacker-Rank 인터뷰 키트 스터디 과제/발표 레포지토리"
