+++ app/src/main/java/org/techtown/soptseminar/SignInActivity.kt
@@ -53,6 +53,7 @@ class SignInActivity : AppCompatActivity() {
 
         }
 
+
         // SignUp
         binding.btnSignup.setOnClickListener() {
             // 회원가입 페이지로 이동!
