+++ app/src/main/res/layout/item_repository_sample_list.xml
@@ -12,7 +12,8 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="10dp"
         android:ellipsize="end"
-        android:maxLines="1"
+        android:maxLines="2"
+        android:paddingEnd="@dimen/margin_10"
         android:text="레포지토리 이름"
         android:textColor="@color/black"
         android:textSize="20sp"
@@ -24,7 +25,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:ellipsize="end"
-        android:maxLines="2"
+        android:maxLines="1"
+        android:paddingEnd="@dimen/margin_10"
         android:text="과제 설명텍스트입니다."
         android:textColor="@color/black"
         android:textSize="14sp"
