+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <dimen name="margin_5">15dp</dimen>
+    <dimen name="margin_10">15dp</dimen>
+    <dimen name="margin_15">15dp</dimen>
+    <dimen name="margin_20">15dp</dimen>
+</resources>
